Improved movement of Shape.
move() and fall() now call a _can_move() method to determine whether or not the
move would be legal. This _can_move() method tests to make sure that the move
or fall would not overrun the screen or overlap another shape. Furthermore,
separating move() from fall() (fall() was previously a convenience method which
simply called move(self, 0, 1) means the user has an extra beat to slide a
piece underneath another on the ground level, which s/he had been unable to do
before.