Added a method to remove finished lines from the board.
This method still needs some work. It should only move boxes ABOVE the line
which has been removed from the board. This becomes slightly more complex if it
is not the bottom line that is removed, and slightly more complex still if
multiple lines are removed. Also, it would be nice if the line flashed or
something. And of course, there is no scoring yet.